Start factoring out a library
